# voice&care-server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Implementação da API Google Assistant"
  version: "1.0.0"
  title: "Voice&Care - HCL"
schemes:
- "http"
- "https"
paths:
  /food:
    get:
      tags:
      - "Conselhos"
      summary: "Receber os conselhos de Alimentação"
      operationId: "foodGET"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Advice"
      x-swagger-router-controller: "Conselhos"
    post:
      tags:
      - "Conselhos"
      summary: "Inserir conselhos de Alimentação"
      operationId: "foodPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "advice"
        required: true
        schema:
          $ref: "#/definitions/Advice"
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Conselhos"
  /rest:
    get:
      tags:
      - "Conselhos"
      summary: "Receber os conselhos de Repouso"
      operationId: "restGET"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Advice"
      x-swagger-router-controller: "Conselhos"
    post:
      tags:
      - "Conselhos"
      summary: "Inserir conselhos de Repouso"
      operationId: "restPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "advice"
        required: true
        schema:
          $ref: "#/definitions/Advice"
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Conselhos"
  /activities:
    get:
      tags:
      - "Conselhos"
      summary: "Receber os conselhos de Atividades"
      operationId: "activitiesGET"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Advice"
      x-swagger-router-controller: "Conselhos"
    post:
      tags:
      - "Conselhos"
      summary: "Inserir conselhos de Atividades"
      operationId: "activitiesPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "advice"
        required: true
        schema:
          $ref: "#/definitions/Advice"
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Conselhos"
  /general:
    get:
      tags:
      - "Conselhos"
      summary: "Receber os conselhos Gerais"
      operationId: "generalGET"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Advice"
      x-swagger-router-controller: "Conselhos"
    post:
      tags:
      - "Conselhos"
      summary: "Inserir conselhos Gerais"
      operationId: "generalPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "advice"
        required: true
        schema:
          $ref: "#/definitions/Advice"
      - name: "user"
        in: "query"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Conselhos"
  /report:
    get:
      tags:
      - "Relatórios"
      summary: "Acessar relatório do paciente"
      operationId: "reportGET"
      produces:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      - name: "datetime"
        in: "query"
        required: true
        type: "string"
        format: "date"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Report"
      x-swagger-router-controller: "Relatrios"
    post:
      tags:
      - "Relatórios"
      summary: "Inserir relatório do paciente"
      operationId: "reportPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "user"
        in: "query"
        required: true
        type: "string"
      - in: "body"
        name: "report"
        required: true
        schema:
          $ref: "#/definitions/Report"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Relatrios"
definitions:
  Report:
    type: "object"
    properties:
      temperature:
        type: "number"
        format: "float"
      bpm:
        type: "integer"
        format: "int32"
      mood:
        type: "integer"
        format: "int32"
      comments:
        type: "string"
    example:
      comments: "comments"
      mood: 1
      temperature: 0.8008282
      bpm: 6
  Advice:
    type: "object"
    required:
    - "text"
    properties:
      text:
        type: "string"
    example:
      text: "text"